Repaired support for mouse events. No support for mouse grabbing yet. Restructured so first level input handling is performed by WindowInputHandler which can feed several ControlSet s to allow switching/multiple sets.

+++ DesktopGUtils/src/com/quew8/gutils/desktop/control/MouseMovementControl.java
@@ -5,5 +5,5 @@ package com.quew8.gutils.desktop.control;
  * @author Quew8
  */
 public class MouseMovementControl {
-    public void movement(int mouseX, int mouseY, int mouseDX, int mouseDY) {}
+    public void movement(double mouseX, double mouseY, double mouseDX, double mouseDY) {}
 }
